+++ src/main/java/faang/school/notificationservice/config/redis/RedisConfig.java
@@ -1,17 +1,16 @@
 package faang.school.notificationservice.config.redis;
 
-import faang.school.notificationservice.listener.AchievementEventListener;
-import faang.school.notificationservice.redis.EventStartEventListener;
+import faang.school.notificationservice.listener.RedisContainerMessageListener;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
 import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 import org.springframework.data.redis.serializer.GenericToStringSerializer;
+
+import java.util.List;
 
 @Configuration
 public class RedisConfig {
@@ -22,9 +21,6 @@ public class RedisConfig {
     @Value("${spring.data.redis.port}")
     private int redisPort;
 
-    @Value("${spring.data.redis.channel.event_start}")
-    private String eventStartTopic;
-
     @Bean
     JedisConnectionFactory jedisConnectionFactory() {
         RedisStandaloneConfiguration redisStandaloneConfiguration =
@@ -34,36 +30,20 @@ public class RedisConfig {
     }
 
     @Bean
-    MessageListenerAdapter achievementMessageListener(AchievementEventListener achievementEventListener) {
-        return new MessageListenerAdapter(achievementEventListener);
-    }
-
-    @Bean
-    public RedisTemplate<String, Object> redisTemplate() {
+    RedisTemplate<String, Object> redisTemplate() {
         final RedisTemplate<String, Object> template = new RedisTemplate<>();
         template.setConnectionFactory(jedisConnectionFactory());
         template.setValueSerializer(new GenericToStringSerializer<>(Object.class));
         return template;
     }
 
-
     @Bean
-    RedisMessageListenerContainer redisContainer(AchievementEventListener achievementEventListener,
-                                                 EventStartEventListener eventStartEventListener) {
+    RedisMessageListenerContainer redisContainer(List<RedisContainerMessageListener> messageListeners) {
         RedisMessageListenerContainer container = new RedisMessageListenerContainer();
         container.setConnectionFactory(jedisConnectionFactory());
-        container.addMessageListener(achievementEventListener.getAdapter(), achievementEventListener.getTopic());
-        container.addMessageListener(eventStartListener(eventStartEventListener), eventStartTopic());
+        messageListeners.forEach(listener ->
+                container.addMessageListener(listener.getAdapter(), listener.getTopic()));
+
         return container;
-    }
-
-    @Bean
-    MessageListenerAdapter eventStartListener(EventStartEventListener eventStartEventListener) {
-        return new MessageListenerAdapter(eventStartEventListener);
-    }
-
-    @Bean
-    ChannelTopic eventStartTopic() {
-        return new ChannelTopic(eventStartTopic);
     }
 }
