+++ src/main/java/faang/school/notificationservice/listener/AchievementEventListener.java
@@ -11,7 +11,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.data.redis.listener.Topic;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -38,11 +37,6 @@ public class AchievementEventListener extends AbstractEventListener<AchievementE
             String messageText = getMessage(event, Locale.ENGLISH);
             sendNotification(event.getUserId(), messageText);
         });
-    }
-
-    @Override
-    public MessageListenerAdapter getAdapter() {
-        return new MessageListenerAdapter(this);
     }
 
     @Override
