+++ src/main/java/faang/school/notificationservice/listener/EventStartEventListener.java
@@ -1,13 +1,14 @@
-package faang.school.notificationservice.redis;
+package faang.school.notificationservice.listener;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.notificationservice.client.UserServiceClient;
 import faang.school.notificationservice.dto.EventStartEventDto;
-import faang.school.notificationservice.listener.AbstractEventListener;
 import faang.school.notificationservice.messaging.MessageBuilder;
 import faang.school.notificationservice.service.NotificationService;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.connection.Message;
-import org.springframework.data.redis.connection.MessageListener;
+import org.springframework.data.redis.listener.ChannelTopic;
+import org.springframework.data.redis.listener.Topic;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -15,7 +16,10 @@ import java.util.Locale;
 import java.util.function.Consumer;
 
 @Component
-public class EventStartEventListener extends AbstractEventListener<EventStartEventDto> implements MessageListener {
+public class EventStartEventListener extends AbstractEventListener<EventStartEventDto> {
+
+    @Value("${spring.data.redis.channel.event_start}")
+    private String eventStartTopic;
 
     public EventStartEventListener(ObjectMapper objectMapper,
                                    UserServiceClient userServiceClient,
@@ -24,12 +28,17 @@ public class EventStartEventListener extends AbstractEventListener<EventStartEve
         super(objectMapper, userServiceClient, notificationService, messageBuilder);
     }
 
-    private Consumer<EventStartEventDto> sendNotifications = eventStartEventDto ->
+    private final Consumer<EventStartEventDto> sendNotifications = eventStartEventDto ->
             eventStartEventDto.getEventParticipants()
                     .forEach(userId -> sendNotification(userId, getMessage(eventStartEventDto, Locale.ENGLISH)));
 
     @Override
     public void onMessage(Message message, byte[] pattern) {
         processEvent(message, EventStartEventDto.class, sendNotifications);
+    }
+
+    @Override
+    public Topic getTopic() {
+        return new ChannelTopic(eventStartTopic);
     }
 }
