+++ src/main/java/faang/school/notificationservice/listener/RedisContainerMessageListener.java
@@ -4,6 +4,10 @@ import org.springframework.data.redis.listener.Topic;
 import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 
 public interface RedisContainerMessageListener {
-    MessageListenerAdapter getAdapter();
+
     Topic getTopic();
+
+    default MessageListenerAdapter getAdapter() {
+        return new MessageListenerAdapter(this);
+    }
 }
